requirements.txt
ipykernel
pandas
scikit-learn
python-dotenv
openai
langchain
pydantic
uvicorn
fastapi
sentence_transformers
langchain-community
boto3